% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{augment}
\alias{augment}
\title{Manipulate a deferred function.}
\usage{
augment(deferred, ...)
}
\arguments{
\item{deferred}{A \code{deferred} function wrapper.}

\item{...}{Name-value pairs, where name is the name of an argument to
the \code{entry} function.}
}
\value{
Modified \code{deferred} function wrapper. Note that the
        original function object is also modified.
}
\description{
Pass a value in place of an argument. This function will
modify the input object.
}
\examples{
d <- defer(function(a, b, c) return(a+b+c))
d <- augment(d, a = 1, b = 2, c = 3)
d()
#> 6

}
