% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defer.R
\name{defer_errors}
\alias{defer_errors}
\title{Run a block of code, collecting deferrable errors}
\usage{
defer_errors(expr)
}
\arguments{
\item{expr}{An expression to evaluate}
}
\description{
Run a block of code, collecting any \code{\link{deferrable_error}}
calls that occur.  Ordinary errors will be thrown immediately.
}
