% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defer.R
\name{deferrable_error}
\alias{deferrable_error}
\title{Create a deferrable error}
\usage{
deferrable_error(message)
}
\arguments{
\item{message}{The error message}
}
\description{
Create an error that will stop immediately, or can be continued from.
}
\examples{
# Deferrable errors will throw immediately if no suitable calling
# handlers are established:
tryCatch(
  deferrable_error("my error"),
  error = identity)

# Create a deferrable error and continue from it, using
# withCallingHandlers:
value <- withCallingHandlers({
   x <- 1
   defer::deferrable_error("a deferrable error")
   x * 2
 },
 deferrable_error = function(e)
   invokeRestart("continue_deferrable_error"))
}
