% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defer.R
\name{deferred_errors_flush}
\alias{deferred_errors_flush}
\title{Flush deferred errors}
\usage{
deferred_errors_flush()
}
\description{
Within a \code{\link{defer_errors}} block, flush any deferred
errors, turning them into realised errors.  If no deferrable
errors have occurred, this function has no effect.
}
\examples{
check_positive <- function(x) {
  if (x < 0) {
    deferrable_error(paste("got a negative number:", x))
  }
}
err <- tryCatch(
  defer::defer_errors({
    check_positive(-1)
    defer::deferred_errors_flush()
    check_positive(-2)
  }),
  error = identity)
err
}
