% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{evaluate_impl.opencpu}
\alias{evaluate_impl.opencpu}
\title{Evaluate with OpenCPU.}
\usage{
\method{evaluate_impl}{opencpu}(ev, package, arguments)
}
\value{
serialized & encoded object and ...
}
\description{
Evaluate with OpenCPU.
}

