% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{extract_functions}
\alias{extract_functions}
\alias{extract_dependencies}
\alias{extract_variables}
\title{\code{list_dependencies} returns a \code{character} vector of
function names that come from other R packages. Each function
name (vector value) has a corresponding package name set in
\code{\link{names}}.}
\usage{
extract_functions(df)

extract_dependencies(df)

extract_variables(df)
}
\arguments{
\item{df}{deferred function object.}
}
\value{
A vector of function names.

\code{list_dependencies} returns a named vector of
functions that belong to other R packages.

`list_dependencies` returns a named list.
}
\description{
\code{list_functions} returns a \code{character} vector
of names of functions packaged in \code{pkg}.

\code{list_dependencies} returns a \code{character} vector of
function names that come from other R packages. Each function
name (vector value) has a corresponding package name set in
\code{\link{names}}.

`list_dependencies` returns a `list` of variables referenced in any
of the packaged functions, that could be found in the caller
environment at the time of packaging.
}
