% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{opencpu}
\alias{opencpu}
\title{Runs a deferred function in a OpenCPU R session.}
\usage{
opencpu(ocpu_url, deferred, ...)
}
\arguments{
\item{ocpu_url}{URL of the OpenCPU server.}

\item{deferred}{A deferred function wrapper.}

\item{...}{Parameters for \code{deferred}, must be named.}
}
\description{
Runs a deferred function in a OpenCPU R session.
}
\examples{
\dontrun{
ocpu_url <- 'https://cloud.opencpu.org/ocpu/'
deferred <- defer(function(x, y)paste(x, y))

opencpu(ocpu_url, deferred, x = 1, y = 2)
}

}
