% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_}
\alias{package_}
\title{Create an execution package.}
\usage{
package_(entry, ..., .funcs)
}
\arguments{
\item{entry}{Entry-point function.}

\item{...}{Other function to be packaged.}

\item{.funcs}{A list of functions.}
}
\value{
An execution package object.
}
\description{
Create an execution package.
}

