% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{print.deferred}
\alias{print.deferred}
\title{Printer for deferred functions.}
\usage{
\method{print}{deferred}(x, ...)
}
\arguments{
\item{x}{object to be printed out.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print.defer} provides a specialized \code{print}
method for deferred function wrappers.
}
