% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{caller_env}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang]{caller_env}}}
}}

