% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{remote_function}
\alias{remote_function}
\title{Deserialize a base64-encoded string and run the deferred function.}
\usage{
remote_function(base64string)
}
\arguments{
\item{base64string}{A serialized, base64-encoded deferred function
wrapper.}
}
\value{
A serialized, base64-encoded result of running the deferred
        wrapper.
}
\description{
The body of this function is used in \code{\link{opencpu}}.
\code{jsonlite} is a dependency of opencpu so for sure it's going
to be there.
}
\details{
This function exists because it's easier to write and test it this
way. However, it is never used directly, but rather deparsed and
sent to a remote server as text.
}
