% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{run_deferred}
\alias{run_deferred}
\title{Run the deferred function.}
\usage{
run_deferred(df, ...)
}
\arguments{
\item{df}{deferred function object.}

\item{...}{parameters passed to \code{df}.}
}
\value{
The return value of the wrapped function.
}
\description{
An explicit way to signify that a function being run is actually
a \code{deferred} function wrapper.
}
