% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{run_package}
\alias{run_package}
\title{Run a function from the execution package.}
\usage{
run_package(pkg, ..., .fun = "entry")
}
\arguments{
\item{pkg}{Execution package object.}

\item{...}{Parameters to the function being called.}

\item{.fun}{Name of the function being called.}
}
\value{
The value returned in the call.
}
\description{
Run a function from the execution package.
}
\seealso{
\code{\link{package_}}
}

