% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{defer}
\alias{defer}
\alias{defer_}
\alias{is_deferred}
\title{Defer function execution.}
\usage{
defer(entry, ..., .dots, .extract = TRUE)

defer_(entry, ..., .dots = list(), .extract = FALSE,
  .caller_env = caller_env(), .verbosity = 1)

is_deferred(x)
}
\arguments{
\item{entry}{Entry-point function.}

\item{...}{List of dependencies, functions and variables.}

\item{.dots}{A list of functions or quosures (see \code{\link[rlang]{quos}}).}

\item{.extract}{Whether to analyze functions and extract dependencies
from their code.}

\item{.caller_env}{The environment where \code{defer_()} is supposed to
assume the call was made and the wrapper is returned to. Its
value is important when \code{.extract} is set to \code{TRUE},
and it is used in the interactive version, \code{defer()}, which
passes its own \code{caller_env()} to \code{defer_()}.}

\item{.verbosity}{Accepts values 0, 1 and 2. 0 means quiet, 1 and 2
result in additional output for the user. Set to \code{1} when
in interactive mode, that is, when called from \code{defer()}.}

\item{x}{Object to be tested.}
}
\value{
A \code{deferred} function object.

\code{TRUE} or \code{FALSE}.
}
\description{
Both \code{defer} and \code{defer_} create an execution package
(wrapper) for any user-provided function.

\code{defer_} is intended for non-interactive use. It
provides an interface very similar to \code{defer} but by default
turns off discovering dependencies (\code{.extract} is \code{FALSE}).

\code{is_deferred} verifies if the given object
is a \code{deferred} function wrapper.
}
\details{
\code{defer} is intended for interactive use - it assumes that
dependencies should be extracted (\code{.extract} defaults to
\code{TRUE}).
}
