% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{parallelize}
\alias{parallelize}
\alias{do_.parallel}
\title{Run grouped \code{do} in parallel.}
\usage{
parallelize(.data)

do_.parallel(.data, ..., env = parent.frame(), .dots = list())
}
\arguments{
\item{.data}{data object transformed with [parallelize]}

\item{...}{expressions to apply to each group; see [dplyr::do]}

\item{env}{the environment in which functions should be evaluated; see [dplyr::do]}

\item{.dots}{pair/values of expressions coercible to lazy objects; see [dplyr::do]}
}
\description{
Injects execution of \code{\link[dplyr]{do}} and if \code{.data} is
groupped, executes the operation in parallel.
}
\examples{
\dontrun{
iris \%>\%
  group_by(Species) \%>\%
  parallelize \%>\%
  do({
    # something that's better run in paralle
    data.frame(result = 1)
  })
}

}
