% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{defer_}
\alias{defer_}
\alias{is_deferred}
\alias{extract_functions}
\alias{extract_dependencies}
\title{Defer function execution - create an execution package.}
\usage{
defer_(entry, ..., .dots = list(), .extract = FALSE)

is_deferred(x)

extract_functions(df)

extract_dependencies(df)
}
\arguments{
\item{entry}{Entry-point function or a function name.}

\item{...}{List of dependencies, functions and variables.}

\item{.extract}{Whether to analyze functions and extract dependencies
from their bodies.}

\item{x}{Object to be tested.}

\item{functions}{A list of functions.}

\item{variables}{A list of variables.}

\item{pkg}{An execution package object.}
}
\value{
A deferred function object.

\code{TRUE} or \code{FALSE}.

A vector of function names.

\code{list_dependencies} returns a named vector of
functions that belong to other R packages.
}
\description{
Defer function execution - create an execution package.

\code{is_deferred} verifies if the given object
is an execution package.

\code{list_functions} returns a \code{character} vector
of names of functions packaged in \code{pkg}.

\code{list_dependencies} returns a \code{character} vector of
function names that come from other R packages. Each function
name (vector value) has a corresponding package name set in
\code{\link{names}}.
}
\seealso{
\code{\link[defer]{execute}}
}
