% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_}
\alias{is_execution_package}
\alias{list_dependencies}
\alias{list_functions}
\alias{package_}
\title{Create an execution package.}
\usage{
package_(entry, ..., .funcs, .extract = FALSE)

is_execution_package(x)

list_functions(pkg)

list_dependencies(pkg)
}
\arguments{
\item{entry}{Entry-point function or a function name.}

\item{...}{Other function to be packaged.}

\item{.funcs}{A list of functions.}

\item{x}{Object to be tested.}

\item{pkg}{An execution package object.}
}
\value{
An execution package object.

\code{TRUE} or \code{FALSE}.

A vector of function names.

\code{list_dependencies} returns a named vector of
functions that belong to other R packages.
}
\description{
If the entry-point function \code{entry} need to be called by a
different name in one of the other functions being packaged,
entry can be set to that name and the function itself can be
passed under its actual name.

\code{is_execution_package} verifies if the given object
is an execution package.

\code{list_functions} returns a \code{character} vector
of names of functions packaged in \code{pkg}.

\code{list_dependencies} returns a \code{character} vector of
function names that come from other R packages. Each function
name (vector value) has a corresponding package name set in
\code{\link{names}}.
}
\seealso{
\code{\link{run_package}}
}

